backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pillow==10.4.0
python-dotenv==1.0.1
sqlite-utils==3.37
easyocr==1.7.2
torch>=2.6.0
torchvision>=0.21.0
opencv-python==4.10.0.84
numpy==1.26.4
